fix: rename Avatar1 to Avatar
